# fastapi-server/openapi.yaml
openapi: 3.0.3
info:
  title: ERCOT Regional Load Data API
  description: API for retrieving aggregated electricity demand data across ERCOT regions
  version: 1.0.0
paths:
  /load/hourly:
    get:
      summary: Get hourly aggregated load data
      description: |
        Retrieves hourly electricity demand data aggregated across all ERCOT regions.
        Returns time-series data showing regional load trends ordered chronologically.
      operationId: getHourlyLoad
      tags:
        - Load Data
      parameters:
        - name: start_date
          in: query
          description: Start date
          required: false
          schema:
            type: string
            format: date-time
            example: "2024-01-01T00:00:00Z"
        - name: end_date
          in: query
          description: End date
          required: false
          schema:
            type: string
            format: date-time
            example: "2024-01-31T23:59:59Z"
        - name: limit
          in: query
          description: Maximum number of records to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 1000
        - name: offset
          in: query
          description: Number of records to skip for pagination
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Successful response with hourly load data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HourlyLoadData'
              examples:
                success:
                  value:
                    - hour_end: "2024-01-01T01:00:00Z"
                      coast: 5432.5
                      east: 8765.2
                      far_west: 2341.8
                      north: 6789.4
                      north_c: 7654.3
                      southern: 9876.1
                      south_c: 8765.4
                      west: 4321.9
                      ercot: 53946.6
                    - hour_end: "2024-01-01T02:00:00Z"
                      coast: 5123.4
                      east: 8234.5
                      far_west: 2198.7
                      north: 6345.2
                      north_c: 7234.5
                      southern: 9345.6
                      south_c: 8234.5
                      west: 4098.3
                      ercot: 50814.7
...
  /weather/heatwaves:
    get:
      summary: Get heatwave streaks by ERCOT zone
      description: |
        Identifies heatwave periods (consecutive days with max temp >= threshold) for ERCOT zones.
        Returns streak start/end dates, duration, and average peak load during each heatwave.
      operationId: getHeatwaveStreaks
      tags:
        - Weather Analysis
      parameters:
        - name: zone
          in: query
          description: Filter by specific ERCOT zone(s). Comma-separated for multiple zones.
          required: false
          schema:
            type: string
            example: "coast,east"
          explode: false
          style: form
        - name: min_temp_f
          in: query
          description: Minimum temperature threshold in Fahrenheit for heatwave definition
          required: false
          schema:
            type: number
            format: double
            default: 100.0
            example: 100.0
        - name: min_days
          in: query
          description: Minimum consecutive days required to qualify as a heatwave
          required: false
          schema:
            type: integer
            minimum: 1
            default: 3
            example: 3
        - name: start_date
          in: query
          description: Filter heatwaves starting on or after this date
          required: false
          schema:
            type: string
            format: date
            example: "2024-01-01"
        - name: end_date
          in: query
          description: Filter heatwaves ending on or before this date
          required: false
          schema:
            type: string
            format: date
            example: "2024-12-31"
      responses:
        '200':
          description: Successful response with heatwave streak data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HeatwaveStreak'
              examples:
                success:
                  value:
                    - zone_code: "coast"
                      streak_start: "2024-07-15"
                      streak_end: "2024-07-19"
                      streak_days: 5
                      max_temp_f: 104.3
                      avg_peak_load_mw: 6543.2
                    - zone_code: "west"
                      streak_start: "2024-08-01"
                      streak_end: "2024-08-04"
                      streak_days: 4
                      max_temp_f: 108.7
                      avg_peak_load_mw: 5234.8
  /weather/precipitation:
    get:
      summary: Compare electricity load on rainy vs. dry days by zone
      description: |
        Analyzes the impact of precipitation on electricity demand by comparing average
        daily load on rainy days versus dry days for each ERCOT zone. A rainy day is
        defined as any day with precipitation > 0mm.
      operationId: getPrecipitationLoadImpact
      tags:
        - Weather Analysis
      parameters:
        - name: zone
          in: query
          description: Filter by specific ERCOT zone(s). Comma-separated for multiple zones.
          required: false
          schema:
            type: string
            example: "coast,north"
          explode: false
          style: form
        - name: start_date
          in: query
          description: Start date for analysis period
          required: false
          schema:
            type: string
            format: date
            example: "2024-01-01"
        - name: end_date
          in: query
          description: End date for analysis period
          required: false
          schema:
            type: string
            format: date
            example: "2024-12-31"
      responses:
        '200':
          description: Successful response with precipitation impact analysis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrecipitationImpact'
              examples:
                success:
                  value:
                    - zone_code: "coast"
                      rainy_day: true
                      avg_load_mw: 5234.8
                      num_days: 87
                    - zone_code: "coast"
                      rainy_day: false
                      avg_load_mw: 5789.3
                      num_days: 278
                    - zone_code: "east"
                      rainy_day: true
                      avg_load_mw: 8012.4
                      num_days: 95
                    - zone_code: "east"
                      rainy_day: false
                      avg_load_mw: 8456.7
                      num_days: 270
  /load/peak-load-extreme-heat:
    get:
      summary: Get median peak load during extreme heat days by zone
      description: |
        Analyzes electricity demand on the hottest days by calculating the median daily
        peak load for extreme heat conditions. Extreme heat is defined by a temperature
        percentile threshold (default: 99th percentile, representing the top 1% hottest days).
        Returns the median peak load, threshold temperature, and number of extreme heat days.
      operationId: getPeakLoadExtremeHeat
      tags:
        - Load Data
      parameters:
        - name: zone
          in: query
          description: Filter by specific ERCOT zone. If omitted, returns data for all zones.
          required: false
          schema:
            type: string
            example: "coast"
            enum:
              - coast
              - east
              - far_west
              - north
              - north_c
              - southern
              - south_c
              - west
        - name: start_date
          in: query
          description: Start date for analysis period (UTC)
          required: false
          schema:
            type: string
            format: date
            example: "2024-01-01"
        - name: end_date
          in: query
          description: End date for analysis period (UTC)
          required: false
          schema:
            type: string
            format: date
            example: "2024-12-31"
        - name: threshold
          in: query
          description: Percentile threshold for defining extreme heat (0-100). Default is 99 (top 1% hottest days).
          required: false
          schema:
            type: number
            format: double
            minimum: 0
            maximum: 100
            default: 99
            example: 99
      responses:
        '200':
          description: Successful response with extreme heat peak load statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExtremeHeatLoad'
              examples:
                success:
                  value:
                    - zone_code: "coast"
                      median_peak_load_mw: 6543.2
                      num_extreme_heat_days: 18
                      threshold_percentile: 99
                      threshold_temp_f: 98.5
                    - zone_code: "east"
                      median_peak_load_mw: 9234.7
                      num_extreme_heat_days: 15
                      threshold_percentile: 99
                      threshold_temp_f: 101.2
                    - zone_code: "west"
                      median_peak_load_mw: 5123.4
                      num_extreme_heat_days: 20
                      threshold_percentile: 99
                      threshold_temp_f: 105.8
...
components:
  schemas:
    HourlyLoadData:
      type: object
      description: Aggregated hourly electricity load data for all ERCOT regions
      properties:
        hour_end:
          type: string
          format: date-time
          description: Timestamp marking the end of the hourly period
          example: "2024-01-01T01:00:00Z"
        coast:
          type: number
          format: double
          description: Total electricity demand for Coast region (MW)
          example: 5432.5
        east:
          type: number
          format: double
          description: Total electricity demand for East region (MW)
          example: 8765.2
        far_west:
          type: number
          format: double
          description: Total electricity demand for Far West region (MW)
          example: 2341.8
        north:
          type: number
          format: double
          description: Total electricity demand for North region (MW)
          example: 6789.4
        north_c:
          type: number
          format: double
          description: Total electricity demand for North Central region (MW)
          example: 7654.3
        southern:
          type: number
          format: double
          description: Total electricity demand for Southern region (MW)
          example: 9876.1
        south_c:
          type: number
          format: double
          description: Total electricity demand for South Central region (MW)
          example: 8765.4
        west:
          type: number
          format: double
          description: Total electricity demand for West region (MW)
          example: 4321.9
        ercot:
          type: number
          format: double
          description: Total electricity demand across entire ERCOT system (MW)
          example: 53946.6
      required:
        - hour_end
        - coast
        - east
        - far_west
        - north
        - north_c
        - southern
        - south_c
        - west
        - ercot
...